+++ optimizer/optimizer_graph_hierarchical.h
@@ -12,17 +12,17 @@
 
 using namespace std;
 
-vector<vector<vector<tuple<float,pair<int,int>,int>>>> compute_partitioning(vector<vector<float>> compute_times,
-                                                                            vector<vector<float>> activation_sizes,
-                                                                            vector<vector<float>> parameter_sizes,
-                                                                            vector<float> output_activation_sizes,
+vector<vector<vector<tuple<double,pair<int,int>,int>>>> compute_partitioning(vector<vector<double>> compute_times,
+                                                                            vector<vector<double>> activation_sizes,
+                                                                            vector<vector<double>> parameter_sizes,
+                                                                            vector<double> output_activation_sizes,
                                                                             vector<vector<int>> all_predecessor_ids,
                                                                             int num_machines, int num_machines_within_machine,
                                                                             int bandwidth, long memory_size,
                                                                             bool straight_pipeline, bool final_level);
-vector<int> analyze_partitioning(vector<vector<vector<tuple<float,pair<int,int>,int>>>> A,
+vector<int> analyze_partitioning(vector<vector<vector<tuple<double,pair<int,int>,int>>>> A,
                                  vector<Node> states, int start, int end,
-                                 int network_bandwidth, int num_machines, float activation_compression_ratio);
+                                 int network_bandwidth, int num_machines, double activation_compression_ratio);
 void optimize_graph(Graph graph, vector<int> all_num_machines, vector<int> network_bandwidths, int memory_size);
 
 #endif //PIPEDREAM_OPTIMIZER_GRAPH_HIERARCHICAL_H
